requirements.txt
bitdotio
pandas
requests
lxml
psycopg2-binary
python-dotenv
sqlalchemy
